streamlit-ai-dental-app/requirements.txt
streamlit
pillow
numpy
scikit-image
reportlab